tg_bot/requirements.txt
aiogram
python-dotenv
pymongo
minio
PyYaml
StrEnum
phonenumbers
pymongo
PyYAML